Finished all the boilerplate for public oplog
