+++ golem-common/src/model/public_oplog/mod.rs
@@ -12,7 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::model::invocation_context::{SpanId, TraceId};
+#[cfg(feature = "protobuf")]
+mod protobuf;
+
+#[cfg(test)]
+mod tests;
+
+use crate::model::invocation_context::{AttributeValue, SpanId, TraceId};
 use crate::model::lucene::{LeafQuery, Query};
 use crate::model::oplog::{DurableFunctionType, LogLevel, OplogIndex, WorkerResourceId};
 use crate::model::plugin::PluginInstallation;
@@ -439,6 +445,16 @@ impl IntoValue for PublicAttributeValue {
     }
 }
 
+impl From<AttributeValue> for PublicAttributeValue {
+    fn from(value: AttributeValue) -> Self {
+        match value {
+            AttributeValue::String(value) => {
+                PublicAttributeValue::String(StringAttributeValue { value })
+            }
+        }
+    }
+}
+
 #[derive(Clone, Debug, Serialize, PartialEq, Deserialize)]
 #[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
 pub struct PublicLocalSpanData {
@@ -509,6 +525,15 @@ pub enum PublicSpanData {
     ExternalSpan(PublicExternalSpanData),
 }
 
+impl PublicSpanData {
+    pub fn span_id(&self) -> &SpanId {
+        match self {
+            PublicSpanData::LocalSpan(data) => &data.span_id,
+            PublicSpanData::ExternalSpan(data) => &data.span_id,
+        }
+    }
+}
+
 impl IntoValue for PublicSpanData {
     fn into_value(self) -> Value {
         match self {
@@ -540,7 +565,7 @@ pub struct ExportedFunctionInvokedParameters {
     pub idempotency_key: IdempotencyKey,
     pub trace_id: TraceId,
     pub trace_states: Vec<String>,
-    pub invocation_context: Vec<PublicSpanData>,
+    pub invocation_context: Vec<Vec<PublicSpanData>>,
 }
 
 impl IntoValue for ExportedFunctionInvokedParameters {
@@ -1866,1446 +1891,5 @@ impl Display for OplogCursor {
             "{}-{}",
             self.next_oplog_index, self.current_component_version
         )
-    }
-}
-
-#[cfg(feature = "protobuf")]
-mod protobuf {
-    use crate::model::oplog::{LogLevel, OplogIndex, WorkerResourceId};
-    use crate::model::public_oplog::{
-        ActivatePluginParameters, CancelInvocationParameters, ChangeRetryPolicyParameters,
-        CreateParameters, DeactivatePluginParameters, DescribeResourceParameters,
-        EndRegionParameters, ErrorParameters, ExportedFunctionCompletedParameters,
-        ExportedFunctionInvokedParameters, ExportedFunctionParameters, FailedUpdateParameters,
-        GrowMemoryParameters, ImportedFunctionInvokedParameters, JumpParameters, LogParameters,
-        ManualUpdateParameters, OplogCursor, PendingUpdateParameters,
-        PendingWorkerInvocationParameters, PluginInstallationDescription,
-        PublicDurableFunctionType, PublicOplogEntry, PublicRetryConfig, PublicUpdateDescription,
-        PublicWorkerInvocation, ResourceParameters, RevertParameters,
-        SnapshotBasedUpdateParameters, SuccessfulUpdateParameters, TimestampParameter,
-        WriteRemoteBatchedParameters,
-    };
-    use crate::model::regions::OplogRegion;
-    use crate::model::Empty;
-    use golem_api_grpc::proto::golem::worker::{
-        oplog_entry, worker_invocation, wrapped_function_type,
-    };
-    use golem_wasm_rpc::ValueAndType;
-    use std::collections::{BTreeMap, BTreeSet, HashMap};
-    use std::time::Duration;
-
-    impl From<golem_api_grpc::proto::golem::worker::OplogCursor> for OplogCursor {
-        fn from(value: golem_api_grpc::proto::golem::worker::OplogCursor) -> Self {
-            Self {
-                next_oplog_index: value.next_oplog_index,
-                current_component_version: value.current_component_version,
-            }
-        }
-    }
-
-    impl From<OplogCursor> for golem_api_grpc::proto::golem::worker::OplogCursor {
-        fn from(value: OplogCursor) -> Self {
-            Self {
-                next_oplog_index: value.next_oplog_index,
-                current_component_version: value.current_component_version,
-            }
-        }
-    }
-
-    impl From<PluginInstallationDescription>
-        for golem_api_grpc::proto::golem::worker::PluginInstallationDescription
-    {
-        fn from(plugin_installation_description: PluginInstallationDescription) -> Self {
-            golem_api_grpc::proto::golem::worker::PluginInstallationDescription {
-                installation_id: Some(plugin_installation_description.installation_id.into()),
-                plugin_name: plugin_installation_description.plugin_name,
-                plugin_version: plugin_installation_description.plugin_version,
-                parameters: HashMap::from_iter(plugin_installation_description.parameters),
-            }
-        }
-    }
-
-    impl TryFrom<golem_api_grpc::proto::golem::worker::PluginInstallationDescription>
-        for PluginInstallationDescription
-    {
-        type Error = String;
-
-        fn try_from(
-            value: golem_api_grpc::proto::golem::worker::PluginInstallationDescription,
-        ) -> Result<Self, Self::Error> {
-            Ok(PluginInstallationDescription {
-                installation_id: value
-                    .installation_id
-                    .ok_or("Missing installation_id".to_string())?
-                    .try_into()?,
-                plugin_name: value.plugin_name,
-                plugin_version: value.plugin_version,
-                parameters: BTreeMap::from_iter(value.parameters),
-            })
-        }
-    }
-    impl TryFrom<golem_api_grpc::proto::golem::worker::OplogEntry> for PublicOplogEntry {
-        type Error = String;
-
-        fn try_from(
-            value: golem_api_grpc::proto::golem::worker::OplogEntry,
-        ) -> Result<Self, String> {
-            match value.entry.ok_or("Oplog entry is empty")? {
-                oplog_entry::Entry::Create(create) => {
-                    Ok(PublicOplogEntry::Create(CreateParameters {
-                        timestamp: create.timestamp.ok_or("Missing timestamp field")?.into(),
-                        worker_id: create
-                            .worker_id
-                            .ok_or("Missing worker_id field")?
-                            .try_into()?,
-                        component_version: create.component_version,
-                        args: create.args,
-                        env: create.env.into_iter().collect(),
-                        account_id: create.account_id.ok_or("Missing account_id field")?.into(),
-                        parent: match create.parent {
-                            Some(parent) => Some(parent.try_into()?),
-                            None => None,
-                        },
-                        component_size: create.component_size,
-                        initial_total_linear_memory_size: create.initial_total_linear_memory_size,
-                        initial_active_plugins: BTreeSet::from_iter(
-                            create
-                                .initial_active_plugins
-                                .into_iter()
-                                .map(|pr| pr.try_into())
-                                .collect::<Result<Vec<_>, _>>()?,
-                        ),
-                    }))
-                }
-                oplog_entry::Entry::ImportedFunctionInvoked(imported_function_invoked) => Ok(
-                    PublicOplogEntry::ImportedFunctionInvoked(ImportedFunctionInvokedParameters {
-                        timestamp: imported_function_invoked
-                            .timestamp
-                            .ok_or("Missing timestamp field")?
-                            .into(),
-                        function_name: imported_function_invoked.function_name,
-                        request: imported_function_invoked
-                            .request
-                            .ok_or("Missing request field")?
-                            .try_into()?,
-                        response: imported_function_invoked
-                            .response
-                            .ok_or("Missing response field")?
-                            .try_into()?,
-                        wrapped_function_type: imported_function_invoked
-                            .wrapped_function_type
-                            .ok_or("Missing wrapped_function_type field")?
-                            .try_into()?,
-                    }),
-                ),
-                oplog_entry::Entry::ExportedFunctionInvoked(exported_function_invoked) => Ok(
-                    PublicOplogEntry::ExportedFunctionInvoked(ExportedFunctionInvokedParameters {
-                        timestamp: exported_function_invoked
-                            .timestamp
-                            .ok_or("Missing timestamp field")?
-                            .into(),
-                        function_name: exported_function_invoked.function_name,
-                        request: exported_function_invoked
-                            .request
-                            .into_iter()
-                            .map(TryInto::try_into)
-                            .collect::<Result<Vec<ValueAndType>, String>>()?,
-                        idempotency_key: exported_function_invoked
-                            .idempotency_key
-                            .ok_or("Missing idempotency_key field")?
-                            .into(),
-                    }),
-                ),
-                oplog_entry::Entry::ExportedFunctionCompleted(exported_function_completed) => {
-                    Ok(PublicOplogEntry::ExportedFunctionCompleted(
-                        ExportedFunctionCompletedParameters {
-                            timestamp: exported_function_completed
-                                .timestamp
-                                .ok_or("Missing timestamp field")?
-                                .into(),
-                            response: exported_function_completed
-                                .response
-                                .ok_or("Missing response field")?
-                                .try_into()?,
-                            consumed_fuel: exported_function_completed.consumed_fuel,
-                        },
-                    ))
-                }
-                oplog_entry::Entry::Suspend(suspend) => {
-                    Ok(PublicOplogEntry::Suspend(TimestampParameter {
-                        timestamp: suspend.timestamp.ok_or("Missing timestamp field")?.into(),
-                    }))
-                }
-                oplog_entry::Entry::Error(error) => Ok(PublicOplogEntry::Error(ErrorParameters {
-                    timestamp: error.timestamp.ok_or("Missing timestamp field")?.into(),
-                    error: error.error,
-                })),
-                oplog_entry::Entry::NoOp(no_op) => Ok(PublicOplogEntry::NoOp(TimestampParameter {
-                    timestamp: no_op.timestamp.ok_or("Missing timestamp field")?.into(),
-                })),
-                oplog_entry::Entry::Jump(jump) => Ok(PublicOplogEntry::Jump(JumpParameters {
-                    timestamp: jump.timestamp.ok_or("Missing timestamp field")?.into(),
-                    jump: OplogRegion {
-                        start: OplogIndex::from_u64(jump.start),
-                        end: OplogIndex::from_u64(jump.end),
-                    },
-                })),
-                oplog_entry::Entry::Interrupted(interrupted) => {
-                    Ok(PublicOplogEntry::Interrupted(TimestampParameter {
-                        timestamp: interrupted
-                            .timestamp
-                            .ok_or("Missing timestamp field")?
-                            .into(),
-                    }))
-                }
-                oplog_entry::Entry::Exited(exited) => {
-                    Ok(PublicOplogEntry::Exited(TimestampParameter {
-                        timestamp: exited.timestamp.ok_or("Missing timestamp field")?.into(),
-                    }))
-                }
-                oplog_entry::Entry::ChangeRetryPolicy(change_retry_policy) => Ok(
-                    PublicOplogEntry::ChangeRetryPolicy(ChangeRetryPolicyParameters {
-                        timestamp: change_retry_policy
-                            .timestamp
-                            .ok_or("Missing timestamp field")?
-                            .into(),
-                        new_policy: change_retry_policy
-                            .retry_policy
-                            .ok_or("Missing retry_policy field")?
-                            .try_into()?,
-                    }),
-                ),
-                oplog_entry::Entry::BeginAtomicRegion(begin_atomic_region) => {
-                    Ok(PublicOplogEntry::BeginAtomicRegion(TimestampParameter {
-                        timestamp: begin_atomic_region
-                            .timestamp
-                            .ok_or("Missing timestamp field")?
-                            .into(),
-                    }))
-                }
-                oplog_entry::Entry::EndAtomicRegion(end_atomic_region) => {
-                    Ok(PublicOplogEntry::EndAtomicRegion(EndRegionParameters {
-                        timestamp: end_atomic_region
-                            .timestamp
-                            .ok_or("Missing timestamp field")?
-                            .into(),
-                        begin_index: OplogIndex::from_u64(end_atomic_region.begin_index),
-                    }))
-                }
-                oplog_entry::Entry::BeginRemoteWrite(begin_remote_write) => {
-                    Ok(PublicOplogEntry::BeginRemoteWrite(TimestampParameter {
-                        timestamp: begin_remote_write
-                            .timestamp
-                            .ok_or("Missing timestamp field")?
-                            .into(),
-                    }))
-                }
-                oplog_entry::Entry::EndRemoteWrite(end_remote_write) => {
-                    Ok(PublicOplogEntry::EndRemoteWrite(EndRegionParameters {
-                        timestamp: end_remote_write
-                            .timestamp
-                            .ok_or("Missing timestamp field")?
-                            .into(),
-                        begin_index: OplogIndex::from_u64(end_remote_write.begin_index),
-                    }))
-                }
-                oplog_entry::Entry::PendingWorkerInvocation(pending_worker_invocation) => Ok(
-                    PublicOplogEntry::PendingWorkerInvocation(PendingWorkerInvocationParameters {
-                        timestamp: pending_worker_invocation
-                            .timestamp
-                            .ok_or("Missing timestamp field")?
-                            .into(),
-                        invocation: pending_worker_invocation
-                            .invocation
-                            .ok_or("Missing invocation field")?
-                            .try_into()?,
-                    }),
-                ),
-                oplog_entry::Entry::PendingUpdate(pending_update) => {
-                    Ok(PublicOplogEntry::PendingUpdate(PendingUpdateParameters {
-                        timestamp: pending_update
-                            .timestamp
-                            .ok_or("Missing timestamp field")?
-                            .into(),
-                        target_version: pending_update.target_version,
-                        description: pending_update
-                            .update_description
-                            .ok_or("Missing update_description field")?
-                            .try_into()?,
-                    }))
-                }
-                oplog_entry::Entry::SuccessfulUpdate(successful_update) => Ok(
-                    PublicOplogEntry::SuccessfulUpdate(SuccessfulUpdateParameters {
-                        timestamp: successful_update
-                            .timestamp
-                            .ok_or("Missing timestamp field")?
-                            .into(),
-                        target_version: successful_update.target_version,
-                        new_component_size: successful_update.new_component_size,
-                        new_active_plugins: BTreeSet::from_iter(
-                            successful_update
-                                .new_active_plugins
-                                .into_iter()
-                                .map(|pr| pr.try_into())
-                                .collect::<Result<Vec<_>, _>>()?,
-                        ),
-                    }),
-                ),
-                oplog_entry::Entry::FailedUpdate(failed_update) => {
-                    Ok(PublicOplogEntry::FailedUpdate(FailedUpdateParameters {
-                        timestamp: failed_update
-                            .timestamp
-                            .ok_or("Missing timestamp field")?
-                            .into(),
-                        target_version: failed_update.target_version,
-                        details: failed_update.details,
-                    }))
-                }
-                oplog_entry::Entry::GrowMemory(grow_memory) => {
-                    Ok(PublicOplogEntry::GrowMemory(GrowMemoryParameters {
-                        timestamp: grow_memory
-                            .timestamp
-                            .ok_or("Missing timestamp field")?
-                            .into(),
-                        delta: grow_memory.delta,
-                    }))
-                }
-                oplog_entry::Entry::CreateResource(create_resource) => {
-                    Ok(PublicOplogEntry::CreateResource(ResourceParameters {
-                        timestamp: create_resource
-                            .timestamp
-                            .ok_or("Missing timestamp field")?
-                            .into(),
-                        id: WorkerResourceId(create_resource.resource_id),
-                    }))
-                }
-                oplog_entry::Entry::DropResource(drop_resource) => {
-                    Ok(PublicOplogEntry::DropResource(ResourceParameters {
-                        timestamp: drop_resource
-                            .timestamp
-                            .ok_or("Missing timestamp field")?
-                            .into(),
-                        id: WorkerResourceId(drop_resource.resource_id),
-                    }))
-                }
-                oplog_entry::Entry::DescribeResource(describe_resource) => Ok(
-                    PublicOplogEntry::DescribeResource(DescribeResourceParameters {
-                        timestamp: describe_resource
-                            .timestamp
-                            .ok_or("Missing timestamp field")?
-                            .into(),
-                        id: WorkerResourceId(describe_resource.resource_id),
-                        resource_name: describe_resource.resource_name,
-                        resource_params: describe_resource
-                            .resource_params
-                            .into_iter()
-                            .map(TryInto::try_into)
-                            .collect::<Result<Vec<ValueAndType>, String>>()?,
-                    }),
-                ),
-                oplog_entry::Entry::Log(log) => Ok(PublicOplogEntry::Log(LogParameters {
-                    level: log.level().into(),
-                    timestamp: log.timestamp.ok_or("Missing timestamp field")?.into(),
-                    context: log.context,
-                    message: log.message,
-                })),
-                oplog_entry::Entry::Restart(restart) => {
-                    Ok(PublicOplogEntry::Restart(TimestampParameter {
-                        timestamp: restart.timestamp.ok_or("Missing timestamp field")?.into(),
-                    }))
-                }
-                oplog_entry::Entry::ActivatePlugin(activate) => {
-                    Ok(PublicOplogEntry::ActivatePlugin(ActivatePluginParameters {
-                        timestamp: activate.timestamp.ok_or("Missing timestamp field")?.into(),
-                        plugin: activate.plugin.ok_or("Missing plugin field")?.try_into()?,
-                    }))
-                }
-                oplog_entry::Entry::DeactivatePlugin(deactivate) => Ok(
-                    PublicOplogEntry::DeactivatePlugin(DeactivatePluginParameters {
-                        timestamp: deactivate
-                            .timestamp
-                            .ok_or("Missing timestamp field")?
-                            .into(),
-                        plugin: deactivate
-                            .plugin
-                            .ok_or("Missing plugin field")?
-                            .try_into()?,
-                    }),
-                ),
-                oplog_entry::Entry::Revert(revert) => {
-                    Ok(PublicOplogEntry::Revert(RevertParameters {
-                        timestamp: revert.timestamp.ok_or("Missing timestamp field")?.into(),
-                        dropped_region: OplogRegion {
-                            start: OplogIndex::from_u64(revert.start),
-                            end: OplogIndex::from_u64(revert.end),
-                        },
-                    }))
-                }
-                oplog_entry::Entry::CancelInvocation(cancel) => Ok(
-                    PublicOplogEntry::CancelInvocation(CancelInvocationParameters {
-                        timestamp: cancel.timestamp.ok_or("Missing timestamp field")?.into(),
-                        idempotency_key: cancel
-                            .idempotency_key
-                            .ok_or("Missing idempotency_key field")?
-                            .into(),
-                    }),
-                ),
-            }
-        }
-    }
-
-    impl TryFrom<PublicOplogEntry> for golem_api_grpc::proto::golem::worker::OplogEntry {
-        type Error = String;
-
-        fn try_from(value: PublicOplogEntry) -> Result<Self, String> {
-            Ok(match value {
-                PublicOplogEntry::Create(create) => golem_api_grpc::proto::golem::worker::OplogEntry {
-                    entry: Some(oplog_entry::Entry::Create(
-                        golem_api_grpc::proto::golem::worker::CreateParameters {
-                            timestamp: Some(create.timestamp.into()),
-                            worker_id: Some(create.worker_id.into()),
-                            component_version: create.component_version,
-                            args: create.args,
-                            env: create.env.into_iter().collect(),
-                            account_id: Some(create.account_id.into()),
-                            parent: create.parent.map(Into::into),
-                            component_size: create.component_size,
-                            initial_total_linear_memory_size: create.initial_total_linear_memory_size,
-                            initial_active_plugins: create
-                                .initial_active_plugins
-                                .into_iter()
-                                .map(Into::into)
-                                .collect(),
-                        },
-                    )),
-                },
-                PublicOplogEntry::ImportedFunctionInvoked(imported_function_invoked) => {
-                    golem_api_grpc::proto::golem::worker::OplogEntry {
-                        entry: Some(oplog_entry::Entry::ImportedFunctionInvoked(
-                            golem_api_grpc::proto::golem::worker::ImportedFunctionInvokedParameters {
-                                timestamp: Some(imported_function_invoked.timestamp.into()),
-                                function_name: imported_function_invoked.function_name.clone(),
-                                request: Some(imported_function_invoked.request.try_into().map_err(
-                                    |errors: Vec<String>| {
-                                        format!("Failed to convert request for imported function {}: {}", imported_function_invoked.function_name, errors.join(", "))
-                                    },
-                                )?),
-                                response: Some(imported_function_invoked.response.try_into().map_err(
-                                    |errors: Vec<String>| {
-                                        format!("Failed to convert response for imported function {}: {}", imported_function_invoked.function_name, errors.join(", "))
-                                    },
-                                )?),
-                                wrapped_function_type: Some(
-                                    imported_function_invoked.wrapped_function_type.into(),
-                                ),
-                            },
-                        )),
-                    }
-                }
-                PublicOplogEntry::ExportedFunctionInvoked(exported_function_invoked) => {
-                    golem_api_grpc::proto::golem::worker::OplogEntry {
-                        entry: Some(oplog_entry::Entry::ExportedFunctionInvoked(
-                            golem_api_grpc::proto::golem::worker::ExportedFunctionInvokedParameters {
-                                timestamp: Some(exported_function_invoked.timestamp.into()),
-                                function_name: exported_function_invoked.function_name.clone(),
-                                request: exported_function_invoked
-                                    .request
-                                    .into_iter()
-                                    .map(|value| {
-                                        value.try_into().map_err(|errors: Vec<String>| {
-                                            format!("Failed to convert request for exported function {}: {}", exported_function_invoked.function_name, errors.join(", "))
-                                        })
-                                    })
-                                    .collect::<Result<Vec<_>, _>>()?,
-                                idempotency_key: Some(exported_function_invoked.idempotency_key.into()),
-                            },
-                        )),
-                    }
-                }
-                PublicOplogEntry::ExportedFunctionCompleted(exported_function_completed) => {
-                    golem_api_grpc::proto::golem::worker::OplogEntry {
-                        entry: Some(oplog_entry::Entry::ExportedFunctionCompleted(
-                            golem_api_grpc::proto::golem::worker::ExportedFunctionCompletedParameters {
-                                timestamp: Some(exported_function_completed.timestamp.into()),
-                                response: Some(
-                                    exported_function_completed.response.try_into().map_err(
-                                        |errors: Vec<String>| {
-                                            format!("Failed to convert response for completed exported function: {}", errors.join(", "))
-                                        },
-                                    )?,
-                                ),
-                                consumed_fuel: exported_function_completed.consumed_fuel,
-                            },
-                        )),
-                    }
-                }
-                PublicOplogEntry::Suspend(suspend) => {
-                    golem_api_grpc::proto::golem::worker::OplogEntry {
-                        entry: Some(oplog_entry::Entry::Suspend(
-                            golem_api_grpc::proto::golem::worker::TimestampParameter {
-                                timestamp: Some(suspend.timestamp.into()),
-                            },
-                        )),
-                    }
-                }
-                PublicOplogEntry::Error(error) => golem_api_grpc::proto::golem::worker::OplogEntry {
-                    entry: Some(oplog_entry::Entry::Error(
-                        golem_api_grpc::proto::golem::worker::ErrorParameters {
-                            timestamp: Some(error.timestamp.into()),
-                            error: error.error,
-                        },
-                    )),
-                },
-                PublicOplogEntry::NoOp(no_op) => golem_api_grpc::proto::golem::worker::OplogEntry {
-                    entry: Some(oplog_entry::Entry::NoOp(
-                        golem_api_grpc::proto::golem::worker::TimestampParameter {
-                            timestamp: Some(no_op.timestamp.into()),
-                        },
-                    )),
-                },
-                PublicOplogEntry::Jump(jump) => golem_api_grpc::proto::golem::worker::OplogEntry {
-                    entry: Some(oplog_entry::Entry::Jump(
-                        golem_api_grpc::proto::golem::worker::JumpParameters {
-                            timestamp: Some(jump.timestamp.into()),
-                            start: jump.jump.start.into(),
-                            end: jump.jump.end.into(),
-                        },
-                    )),
-                },
-                PublicOplogEntry::Interrupted(interrupted) => {
-                    golem_api_grpc::proto::golem::worker::OplogEntry {
-                        entry: Some(oplog_entry::Entry::Interrupted(
-                            golem_api_grpc::proto::golem::worker::TimestampParameter {
-                                timestamp: Some(interrupted.timestamp.into()),
-                            },
-                        )),
-                    }
-                }
-                PublicOplogEntry::Exited(exited) => golem_api_grpc::proto::golem::worker::OplogEntry {
-                    entry: Some(oplog_entry::Entry::Exited(
-                        golem_api_grpc::proto::golem::worker::TimestampParameter {
-                            timestamp: Some(exited.timestamp.into()),
-                        },
-                    )),
-                },
-                PublicOplogEntry::ChangeRetryPolicy(change_retry_policy) => {
-                    golem_api_grpc::proto::golem::worker::OplogEntry {
-                        entry: Some(oplog_entry::Entry::ChangeRetryPolicy(
-                            golem_api_grpc::proto::golem::worker::ChangeRetryPolicyParameters {
-                                timestamp: Some(change_retry_policy.timestamp.into()),
-                                retry_policy: Some(change_retry_policy.new_policy.into()),
-                            },
-                        )),
-                    }
-                }
-                PublicOplogEntry::BeginAtomicRegion(begin_atomic_region) => {
-                    golem_api_grpc::proto::golem::worker::OplogEntry {
-                        entry: Some(oplog_entry::Entry::BeginAtomicRegion(
-                            golem_api_grpc::proto::golem::worker::TimestampParameter {
-                                timestamp: Some(begin_atomic_region.timestamp.into()),
-                            },
-                        )),
-                    }
-                }
-                PublicOplogEntry::EndAtomicRegion(end_atomic_region) => {
-                    golem_api_grpc::proto::golem::worker::OplogEntry {
-                        entry: Some(oplog_entry::Entry::EndAtomicRegion(
-                            golem_api_grpc::proto::golem::worker::EndAtomicRegionParameters {
-                                timestamp: Some(end_atomic_region.timestamp.into()),
-                                begin_index: end_atomic_region.begin_index.into(),
-                            },
-                        )),
-                    }
-                }
-                PublicOplogEntry::BeginRemoteWrite(begin_remote_write) => {
-                    golem_api_grpc::proto::golem::worker::OplogEntry {
-                        entry: Some(oplog_entry::Entry::BeginRemoteWrite(
-                            golem_api_grpc::proto::golem::worker::TimestampParameter {
-                                timestamp: Some(begin_remote_write.timestamp.into()),
-                            },
-                        )),
-                    }
-                }
-                PublicOplogEntry::EndRemoteWrite(end_remote_write) => {
-                    golem_api_grpc::proto::golem::worker::OplogEntry {
-                        entry: Some(oplog_entry::Entry::EndRemoteWrite(
-                            golem_api_grpc::proto::golem::worker::EndRemoteWriteParameters {
-                                timestamp: Some(end_remote_write.timestamp.into()),
-                                begin_index: end_remote_write.begin_index.into(),
-                            },
-                        )),
-                    }
-                }
-                PublicOplogEntry::PendingWorkerInvocation(pending_worker_invocation) => {
-                    golem_api_grpc::proto::golem::worker::OplogEntry {
-                        entry: Some(oplog_entry::Entry::PendingWorkerInvocation(
-                            golem_api_grpc::proto::golem::worker::PendingWorkerInvocationParameters {
-                                timestamp: Some(pending_worker_invocation.timestamp.into()),
-                                invocation: Some(pending_worker_invocation.invocation.try_into()?),
-                            },
-                        )),
-                    }
-                }
-                PublicOplogEntry::PendingUpdate(pending_update) => {
-                    golem_api_grpc::proto::golem::worker::OplogEntry {
-                        entry: Some(oplog_entry::Entry::PendingUpdate(
-                            golem_api_grpc::proto::golem::worker::PendingUpdateParameters {
-                                timestamp: Some(pending_update.timestamp.into()),
-                                target_version: pending_update.target_version,
-                                update_description: Some(pending_update.description.into()),
-                            },
-                        )),
-                    }
-                }
-                PublicOplogEntry::SuccessfulUpdate(successful_update) => {
-                    golem_api_grpc::proto::golem::worker::OplogEntry {
-                        entry: Some(oplog_entry::Entry::SuccessfulUpdate(
-                            golem_api_grpc::proto::golem::worker::SuccessfulUpdateParameters {
-                                timestamp: Some(successful_update.timestamp.into()),
-                                target_version: successful_update.target_version,
-                                new_component_size: successful_update.new_component_size,
-                                new_active_plugins: successful_update
-                                    .new_active_plugins
-                                    .into_iter()
-                                    .map(Into::into)
-                                    .collect(),
-                            },
-                        )),
-                    }
-                }
-                PublicOplogEntry::FailedUpdate(failed_update) => {
-                    golem_api_grpc::proto::golem::worker::OplogEntry {
-                        entry: Some(oplog_entry::Entry::FailedUpdate(
-                            golem_api_grpc::proto::golem::worker::FailedUpdateParameters {
-                                timestamp: Some(failed_update.timestamp.into()),
-                                target_version: failed_update.target_version,
-                                details: failed_update.details,
-                            },
-                        )),
-                    }
-                }
-                PublicOplogEntry::GrowMemory(grow_memory) => {
-                    golem_api_grpc::proto::golem::worker::OplogEntry {
-                        entry: Some(oplog_entry::Entry::GrowMemory(
-                            golem_api_grpc::proto::golem::worker::GrowMemoryParameters {
-                                timestamp: Some(grow_memory.timestamp.into()),
-                                delta: grow_memory.delta,
-                            },
-                        )),
-                    }
-                }
-                PublicOplogEntry::CreateResource(create_resource) => {
-                    golem_api_grpc::proto::golem::worker::OplogEntry {
-                        entry: Some(oplog_entry::Entry::CreateResource(
-                            golem_api_grpc::proto::golem::worker::CreateResourceParameters {
-                                timestamp: Some(create_resource.timestamp.into()),
-                                resource_id: create_resource.id.0,
-                            },
-                        )),
-                    }
-                }
-                PublicOplogEntry::DropResource(drop_resource) => {
-                    golem_api_grpc::proto::golem::worker::OplogEntry {
-                        entry: Some(oplog_entry::Entry::DropResource(
-                            golem_api_grpc::proto::golem::worker::DropResourceParameters {
-                                timestamp: Some(drop_resource.timestamp.into()),
-                                resource_id: drop_resource.id.0,
-                            },
-                        )),
-                    }
-                }
-                PublicOplogEntry::DescribeResource(describe_resource) => {
-                    golem_api_grpc::proto::golem::worker::OplogEntry {
-                        entry: Some(oplog_entry::Entry::DescribeResource(
-                            golem_api_grpc::proto::golem::worker::DescribeResourceParameters {
-                                timestamp: Some(describe_resource.timestamp.into()),
-                                resource_id: describe_resource.id.0,
-                                resource_name: describe_resource.resource_name,
-                                resource_params: describe_resource
-                                    .resource_params
-                                    .into_iter()
-                                    .map(|value| {
-                                        value.try_into().map_err(|errors: Vec<String>| {
-                                            format!("Failed to convert request: {}", errors.join(", "))
-                                        })
-                                    })
-                                    .collect::<Result<Vec<_>, _>>()?,
-                            },
-                        )),
-                    }
-                }
-                PublicOplogEntry::Log(log) => golem_api_grpc::proto::golem::worker::OplogEntry {
-                    entry: Some(oplog_entry::Entry::Log(
-                        golem_api_grpc::proto::golem::worker::LogParameters {
-                            timestamp: Some(log.timestamp.into()),
-                            level: Into::<golem_api_grpc::proto::golem::worker::OplogLogLevel>::into(
-                                log.level,
-                            ) as i32,
-                            context: log.context,
-                            message: log.message,
-                        },
-                    )),
-                },
-                PublicOplogEntry::Restart(restart) => {
-                    golem_api_grpc::proto::golem::worker::OplogEntry {
-                        entry: Some(oplog_entry::Entry::Restart(
-                            golem_api_grpc::proto::golem::worker::TimestampParameter {
-                                timestamp: Some(restart.timestamp.into()),
-                            },
-                        )),
-                    }
-                }
-                PublicOplogEntry::ActivatePlugin(activate) => {
-                    golem_api_grpc::proto::golem::worker::OplogEntry {
-                        entry: Some(oplog_entry::Entry::ActivatePlugin(
-                            golem_api_grpc::proto::golem::worker::ActivatePluginParameters {
-                                timestamp: Some(activate.timestamp.into()),
-                                plugin: Some(activate.plugin.into()),
-                            },
-                        )),
-                    }
-                }
-                PublicOplogEntry::DeactivatePlugin(deactivate) => {
-                    golem_api_grpc::proto::golem::worker::OplogEntry {
-                        entry: Some(oplog_entry::Entry::DeactivatePlugin(
-                            golem_api_grpc::proto::golem::worker::DeactivatePluginParameters {
-                                timestamp: Some(deactivate.timestamp.into()),
-                                plugin: Some(deactivate.plugin.into()),
-                            },
-                        )),
-                    }
-                }
-                PublicOplogEntry::Revert(revert) => {
-                    golem_api_grpc::proto::golem::worker::OplogEntry {
-                        entry: Some(oplog_entry::Entry::Revert(
-                            golem_api_grpc::proto::golem::worker::RevertParameters {
-                                timestamp: Some(revert.timestamp.into()),
-                                start: revert.dropped_region.start.0,
-                                end: revert.dropped_region.end.0,
-                            },
-                        )),
-                    }
-                }
-                PublicOplogEntry::CancelInvocation(cancel) => {
-                    golem_api_grpc::proto::golem::worker::OplogEntry {
-                        entry: Some(oplog_entry::Entry::CancelInvocation(
-                            golem_api_grpc::proto::golem::worker::CancelInvocationParameters {
-                                timestamp: Some(cancel.timestamp.into()),
-                                idempotency_key: Some(cancel.idempotency_key.into()),
-                            },
-                        )),
-                    }
-                }
-            })
-        }
-    }
-
-    impl TryFrom<golem_api_grpc::proto::golem::worker::WrappedFunctionType>
-        for PublicDurableFunctionType
-    {
-        type Error = String;
-
-        fn try_from(
-            value: golem_api_grpc::proto::golem::worker::WrappedFunctionType,
-        ) -> Result<Self, Self::Error> {
-            match value.r#type() {
-                wrapped_function_type::Type::ReadLocal => {
-                    Ok(PublicDurableFunctionType::ReadLocal(Empty {}))
-                }
-                wrapped_function_type::Type::WriteLocal => {
-                    Ok(PublicDurableFunctionType::WriteLocal(Empty {}))
-                }
-                wrapped_function_type::Type::ReadRemote => {
-                    Ok(PublicDurableFunctionType::ReadRemote(Empty {}))
-                }
-                wrapped_function_type::Type::WriteRemote => {
-                    Ok(PublicDurableFunctionType::WriteRemote(Empty {}))
-                }
-                wrapped_function_type::Type::WriteRemoteBatched => Ok(
-                    PublicDurableFunctionType::WriteRemoteBatched(WriteRemoteBatchedParameters {
-                        index: value.oplog_index.map(OplogIndex::from_u64),
-                    }),
-                ),
-            }
-        }
-    }
-
-    impl From<PublicDurableFunctionType> for golem_api_grpc::proto::golem::worker::WrappedFunctionType {
-        fn from(value: PublicDurableFunctionType) -> Self {
-            match value {
-                PublicDurableFunctionType::ReadLocal(_) => {
-                    golem_api_grpc::proto::golem::worker::WrappedFunctionType {
-                        r#type: wrapped_function_type::Type::ReadLocal as i32,
-                        oplog_index: None,
-                    }
-                }
-                PublicDurableFunctionType::WriteLocal(_) => {
-                    golem_api_grpc::proto::golem::worker::WrappedFunctionType {
-                        r#type: wrapped_function_type::Type::WriteLocal as i32,
-                        oplog_index: None,
-                    }
-                }
-                PublicDurableFunctionType::ReadRemote(_) => {
-                    golem_api_grpc::proto::golem::worker::WrappedFunctionType {
-                        r#type: wrapped_function_type::Type::ReadRemote as i32,
-                        oplog_index: None,
-                    }
-                }
-                PublicDurableFunctionType::WriteRemote(_) => {
-                    golem_api_grpc::proto::golem::worker::WrappedFunctionType {
-                        r#type: wrapped_function_type::Type::WriteRemote as i32,
-                        oplog_index: None,
-                    }
-                }
-                PublicDurableFunctionType::WriteRemoteBatched(parameters) => {
-                    golem_api_grpc::proto::golem::worker::WrappedFunctionType {
-                        r#type: wrapped_function_type::Type::WriteRemoteBatched as i32,
-                        oplog_index: parameters.index.map(|index| index.into()),
-                    }
-                }
-            }
-        }
-    }
-
-    impl TryFrom<golem_api_grpc::proto::golem::worker::RetryPolicy> for PublicRetryConfig {
-        type Error = String;
-
-        fn try_from(
-            value: golem_api_grpc::proto::golem::worker::RetryPolicy,
-        ) -> Result<Self, Self::Error> {
-            Ok(PublicRetryConfig {
-                max_attempts: value.max_attempts,
-                min_delay: Duration::from_millis(value.min_delay),
-                max_delay: Duration::from_millis(value.max_delay),
-                multiplier: value.multiplier,
-                max_jitter_factor: value.max_jitter_factor,
-            })
-        }
-    }
-
-    impl From<PublicRetryConfig> for golem_api_grpc::proto::golem::worker::RetryPolicy {
-        fn from(value: PublicRetryConfig) -> Self {
-            golem_api_grpc::proto::golem::worker::RetryPolicy {
-                max_attempts: value.max_attempts,
-                min_delay: value.min_delay.as_millis() as u64,
-                max_delay: value.max_delay.as_millis() as u64,
-                multiplier: value.multiplier,
-                max_jitter_factor: value.max_jitter_factor,
-            }
-        }
-    }
-
-    impl From<golem_api_grpc::proto::golem::worker::OplogLogLevel> for LogLevel {
-        fn from(value: golem_api_grpc::proto::golem::worker::OplogLogLevel) -> Self {
-            match value {
-                golem_api_grpc::proto::golem::worker::OplogLogLevel::OplogTrace => LogLevel::Trace,
-                golem_api_grpc::proto::golem::worker::OplogLogLevel::OplogDebug => LogLevel::Debug,
-                golem_api_grpc::proto::golem::worker::OplogLogLevel::OplogInfo => LogLevel::Info,
-                golem_api_grpc::proto::golem::worker::OplogLogLevel::OplogWarn => LogLevel::Warn,
-                golem_api_grpc::proto::golem::worker::OplogLogLevel::OplogError => LogLevel::Error,
-                golem_api_grpc::proto::golem::worker::OplogLogLevel::OplogCritical => {
-                    LogLevel::Critical
-                }
-                golem_api_grpc::proto::golem::worker::OplogLogLevel::OplogStderr => {
-                    LogLevel::Stderr
-                }
-                golem_api_grpc::proto::golem::worker::OplogLogLevel::OplogStdout => {
-                    LogLevel::Stdout
-                }
-            }
-        }
-    }
-
-    impl From<LogLevel> for golem_api_grpc::proto::golem::worker::OplogLogLevel {
-        fn from(value: LogLevel) -> Self {
-            match value {
-                LogLevel::Trace => golem_api_grpc::proto::golem::worker::OplogLogLevel::OplogTrace,
-                LogLevel::Debug => golem_api_grpc::proto::golem::worker::OplogLogLevel::OplogDebug,
-                LogLevel::Info => golem_api_grpc::proto::golem::worker::OplogLogLevel::OplogInfo,
-                LogLevel::Warn => golem_api_grpc::proto::golem::worker::OplogLogLevel::OplogWarn,
-                LogLevel::Error => golem_api_grpc::proto::golem::worker::OplogLogLevel::OplogError,
-                LogLevel::Critical => {
-                    golem_api_grpc::proto::golem::worker::OplogLogLevel::OplogCritical
-                }
-                LogLevel::Stderr => {
-                    golem_api_grpc::proto::golem::worker::OplogLogLevel::OplogStderr
-                }
-                LogLevel::Stdout => {
-                    golem_api_grpc::proto::golem::worker::OplogLogLevel::OplogStdout
-                }
-            }
-        }
-    }
-
-    impl TryFrom<golem_api_grpc::proto::golem::worker::WorkerInvocation> for PublicWorkerInvocation {
-        type Error = String;
-
-        fn try_from(
-            value: golem_api_grpc::proto::golem::worker::WorkerInvocation,
-        ) -> Result<Self, Self::Error> {
-            match value.invocation.ok_or("Missing invocation field")? {
-                worker_invocation::Invocation::ExportedFunction(exported_function) => Ok(
-                    PublicWorkerInvocation::ExportedFunction(ExportedFunctionParameters {
-                        idempotency_key: exported_function
-                            .idempotency_key
-                            .ok_or("Missing idempotency_key field")?
-                            .into(),
-                        full_function_name: exported_function.function_name,
-                        function_input: if exported_function.valid_input {
-                            Some(
-                                exported_function
-                                    .input
-                                    .into_iter()
-                                    .map(TryInto::try_into)
-                                    .collect::<Result<Vec<ValueAndType>, String>>()?,
-                            )
-                        } else {
-                            None
-                        },
-                    }),
-                ),
-                worker_invocation::Invocation::ManualUpdate(manual_update) => Ok(
-                    PublicWorkerInvocation::ManualUpdate(ManualUpdateParameters {
-                        target_version: manual_update,
-                    }),
-                ),
-            }
-        }
-    }
-
-    impl TryFrom<PublicWorkerInvocation> for golem_api_grpc::proto::golem::worker::WorkerInvocation {
-        type Error = String;
-
-        fn try_from(value: PublicWorkerInvocation) -> Result<Self, Self::Error> {
-            Ok(match value {
-                PublicWorkerInvocation::ExportedFunction(exported_function) => {
-                    golem_api_grpc::proto::golem::worker::WorkerInvocation {
-                        invocation: Some(worker_invocation::Invocation::ExportedFunction(
-                            golem_api_grpc::proto::golem::worker::ExportedFunctionInvocationParameters {
-                                idempotency_key: Some(exported_function.idempotency_key.into()),
-                                function_name: exported_function.full_function_name,
-                                valid_input: exported_function.function_input.is_some(),
-                                input: exported_function
-                                    .function_input
-                                    .unwrap_or_default()
-                                    .into_iter()
-                                    .map(|input| input.try_into().map_err(
-                                        |errors: Vec<String>| {
-                                            format!("Failed to convert request: {}", errors.join(", "))
-                                        },
-                                    )).collect::<Result<Vec<_>, _>>()?,
-                            },
-                        )),
-                    }
-                }
-                PublicWorkerInvocation::ManualUpdate(manual_update) => {
-                    golem_api_grpc::proto::golem::worker::WorkerInvocation {
-                        invocation: Some(worker_invocation::Invocation::ManualUpdate(
-                            manual_update.target_version,
-                        )),
-                    }
-                }
-            })
-        }
-    }
-
-    impl TryFrom<golem_api_grpc::proto::golem::worker::UpdateDescription> for PublicUpdateDescription {
-        type Error = String;
-
-        fn try_from(
-            value: golem_api_grpc::proto::golem::worker::UpdateDescription,
-        ) -> Result<Self, Self::Error> {
-            match value.description.ok_or("Missing description field")? {
-                golem_api_grpc::proto::golem::worker::update_description::Description::AutoUpdate(_) => {
-                    Ok(PublicUpdateDescription::Automatic(Empty {}))
-                }
-                golem_api_grpc::proto::golem::worker::update_description::Description::SnapshotBased(
-                    snapshot_based,
-                ) => Ok(PublicUpdateDescription::SnapshotBased(SnapshotBasedUpdateParameters {
-                    payload: snapshot_based.payload,
-                })),
-            }
-        }
-    }
-
-    impl From<PublicUpdateDescription> for golem_api_grpc::proto::golem::worker::UpdateDescription {
-        fn from(value: PublicUpdateDescription) -> Self {
-            match value {
-                PublicUpdateDescription::Automatic(_) => golem_api_grpc::proto::golem::worker::UpdateDescription {
-                    description: Some(
-                        golem_api_grpc::proto::golem::worker::update_description::Description::AutoUpdate(
-                            golem_api_grpc::proto::golem::common::Empty {},
-                        ),
-                    ),
-                },
-                PublicUpdateDescription::SnapshotBased(snapshot_based) => {
-                    golem_api_grpc::proto::golem::worker::UpdateDescription {
-                        description: Some(
-                            golem_api_grpc::proto::golem::worker::update_description::Description::SnapshotBased(
-                                golem_api_grpc::proto::golem::worker::SnapshotBasedUpdateParameters {
-                                    payload: snapshot_based.payload
-                                }
-                            ),
-                        ),
-                    }
-                }
-            }
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use test_r::test;
-
-    use crate::model::public_oplog::{
-        ChangeRetryPolicyParameters, CreateParameters, DescribeResourceParameters,
-        EndRegionParameters, ErrorParameters, ExportedFunctionCompletedParameters,
-        ExportedFunctionInvokedParameters, ExportedFunctionParameters, FailedUpdateParameters,
-        GrowMemoryParameters, ImportedFunctionInvokedParameters, JumpParameters, LogParameters,
-        PendingUpdateParameters, PendingWorkerInvocationParameters, PluginInstallationDescription,
-        PublicDurableFunctionType, PublicOplogEntry, PublicRetryConfig, PublicUpdateDescription,
-        PublicWorkerInvocation, ResourceParameters, SnapshotBasedUpdateParameters,
-        SuccessfulUpdateParameters, TimestampParameter,
-    };
-    use crate::model::{
-        AccountId, ComponentId, Empty, IdempotencyKey, PluginInstallationId, Timestamp, WorkerId,
-    };
-    use std::collections::{BTreeMap, BTreeSet};
-    use uuid::Uuid;
-
-    use crate::model::oplog::{LogLevel, OplogIndex, WorkerResourceId};
-    use crate::model::regions::OplogRegion;
-    use golem_wasm_ast::analysis::analysed_type::{field, list, r#enum, record, s16, str, u64};
-    use golem_wasm_rpc::{Value, ValueAndType};
-    #[cfg(feature = "poem")]
-    use poem_openapi::types::ToJSON;
-
-    fn rounded_ts(ts: Timestamp) -> Timestamp {
-        Timestamp::from(ts.to_millis())
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn create_serialization_poem_serde_equivalence() {
-        let entry = PublicOplogEntry::Create(CreateParameters {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-            worker_id: WorkerId {
-                component_id: ComponentId(
-                    Uuid::parse_str("13A5C8D4-F05E-4E23-B982-F4D413E181CB").unwrap(),
-                ),
-                worker_name: "test1".to_string(),
-            },
-            component_version: 1,
-            args: vec!["a".to_string(), "b".to_string()],
-            env: vec![("x".to_string(), "y".to_string())]
-                .into_iter()
-                .collect(),
-            account_id: AccountId {
-                value: "account_id".to_string(),
-            },
-            parent: Some(WorkerId {
-                component_id: ComponentId(
-                    Uuid::parse_str("13A5C8D4-F05E-4E23-B982-F4D413E181CB").unwrap(),
-                ),
-                worker_name: "test2".to_string(),
-            }),
-            component_size: 100_000_000,
-            initial_total_linear_memory_size: 200_000_000,
-            initial_active_plugins: BTreeSet::from_iter(vec![PluginInstallationDescription {
-                installation_id: PluginInstallationId(
-                    Uuid::parse_str("13A5C8D4-F05E-4E23-B982-F4D413E181CB").unwrap(),
-                ),
-                plugin_name: "plugin1".to_string(),
-                plugin_version: "1".to_string(),
-                parameters: BTreeMap::new(),
-            }]),
-        });
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn imported_function_invoked_serialization_poem_serde_equivalence() {
-        let entry = PublicOplogEntry::ImportedFunctionInvoked(ImportedFunctionInvokedParameters {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-            function_name: "test".to_string(),
-            request: ValueAndType {
-                value: Value::String("test".to_string()),
-                typ: str(),
-            },
-            response: ValueAndType {
-                value: Value::List(vec![Value::U64(1)]),
-                typ: list(u64()),
-            },
-            wrapped_function_type: PublicDurableFunctionType::ReadRemote(Empty {}),
-        });
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn exported_function_invoked_serialization_poem_serde_equivalence() {
-        let entry = PublicOplogEntry::ExportedFunctionInvoked(ExportedFunctionInvokedParameters {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-            function_name: "test".to_string(),
-            request: vec![
-                ValueAndType {
-                    value: Value::String("test".to_string()),
-                    typ: str(),
-                },
-                ValueAndType {
-                    value: Value::Record(vec![Value::S16(1), Value::S16(-1)]),
-                    typ: record(vec![field("x", s16()), field("y", s16())]),
-                },
-            ],
-            idempotency_key: IdempotencyKey::new("idempotency_key".to_string()),
-        });
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn exported_function_completed_serialization_poem_serde_equivalence() {
-        let entry =
-            PublicOplogEntry::ExportedFunctionCompleted(ExportedFunctionCompletedParameters {
-                timestamp: rounded_ts(Timestamp::now_utc()),
-                response: ValueAndType {
-                    value: Value::Enum(1),
-                    typ: r#enum(&["red", "green", "blue"]),
-                },
-                consumed_fuel: 100,
-            });
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn suspend_serialization_poem_serde_equivalence() {
-        let entry = PublicOplogEntry::Suspend(TimestampParameter {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-        });
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn error_serialization_poem_serde_equivalence() {
-        let entry = PublicOplogEntry::Error(ErrorParameters {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-            error: "test".to_string(),
-        });
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn no_op_serialization_poem_serde_equivalence() {
-        let entry = PublicOplogEntry::NoOp(TimestampParameter {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-        });
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn jump_serialization_poem_serde_equivalence() {
-        let entry = PublicOplogEntry::Jump(JumpParameters {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-            jump: OplogRegion {
-                start: OplogIndex::from_u64(1),
-                end: OplogIndex::from_u64(2),
-            },
-        });
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn interrupted_serialization_poem_serde_equivalence() {
-        let entry = PublicOplogEntry::Interrupted(TimestampParameter {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-        });
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn exited_serialization_poem_serde_equivalence() {
-        let entry = PublicOplogEntry::Exited(TimestampParameter {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-        });
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn change_retry_policy_serialization_poem_serde_equivalence() {
-        let entry = PublicOplogEntry::ChangeRetryPolicy(ChangeRetryPolicyParameters {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-            new_policy: PublicRetryConfig {
-                max_attempts: 10,
-                min_delay: std::time::Duration::from_secs(1),
-                max_delay: std::time::Duration::from_secs(10),
-                multiplier: 2.0,
-                max_jitter_factor: Some(0.1),
-            },
-        });
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn begin_atomic_region_serialization_poem_serde_equivalence() {
-        let entry = PublicOplogEntry::BeginAtomicRegion(TimestampParameter {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-        });
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn end_atomic_region_serialization_poem_serde_equivalence() {
-        let entry = PublicOplogEntry::EndAtomicRegion(EndRegionParameters {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-            begin_index: OplogIndex::from_u64(1),
-        });
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn begin_remote_write_serialization_poem_serde_equivalence() {
-        let entry = PublicOplogEntry::BeginRemoteWrite(TimestampParameter {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-        });
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn end_remote_write_serialization_poem_serde_equivalence() {
-        let entry = PublicOplogEntry::EndRemoteWrite(EndRegionParameters {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-            begin_index: OplogIndex::from_u64(1),
-        });
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn pending_worker_invocation_serialization_poem_serde_equivalence() {
-        let entry = PublicOplogEntry::PendingWorkerInvocation(PendingWorkerInvocationParameters {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-            invocation: PublicWorkerInvocation::ExportedFunction(ExportedFunctionParameters {
-                idempotency_key: IdempotencyKey::new("idempotency_key".to_string()),
-                full_function_name: "test".to_string(),
-                function_input: Some(vec![
-                    ValueAndType {
-                        value: Value::String("test".to_string()),
-                        typ: str(),
-                    },
-                    ValueAndType {
-                        value: Value::Record(vec![Value::S16(1), Value::S16(-1)]),
-                        typ: record(vec![field("x", s16()), field("y", s16())]),
-                    },
-                ]),
-            }),
-        });
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn pending_update_serialization_poem_serde_equivalence_1() {
-        let entry = PublicOplogEntry::PendingUpdate(PendingUpdateParameters {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-            target_version: 1,
-            description: PublicUpdateDescription::SnapshotBased(SnapshotBasedUpdateParameters {
-                payload: "test".as_bytes().to_vec(),
-            }),
-        });
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn pending_update_serialization_poem_serde_equivalence_2() {
-        let entry = PublicOplogEntry::PendingUpdate(PendingUpdateParameters {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-            target_version: 1,
-            description: PublicUpdateDescription::Automatic(Empty {}),
-        });
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn successful_update_serialization_poem_serde_equivalence() {
-        let entry = PublicOplogEntry::SuccessfulUpdate(SuccessfulUpdateParameters {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-            target_version: 1,
-            new_component_size: 100_000_000,
-            new_active_plugins: BTreeSet::from_iter(vec![PluginInstallationDescription {
-                installation_id: PluginInstallationId(
-                    Uuid::parse_str("13A5C8D4-F05E-4E23-B982-F4D413E181CB").unwrap(),
-                ),
-                plugin_name: "plugin1".to_string(),
-                plugin_version: "1".to_string(),
-                parameters: BTreeMap::new(),
-            }]),
-        });
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn failed_update_serialization_poem_serde_equivalence_1() {
-        let entry = PublicOplogEntry::FailedUpdate(FailedUpdateParameters {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-            target_version: 1,
-            details: Some("test".to_string()),
-        });
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn failed_update_serialization_poem_serde_equivalence_2() {
-        let entry = PublicOplogEntry::FailedUpdate(FailedUpdateParameters {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-            target_version: 1,
-            details: None,
-        });
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn grow_memory_serialization_poem_serde_equivalence() {
-        let entry = PublicOplogEntry::GrowMemory(GrowMemoryParameters {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-            delta: 100_000_000,
-        });
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn create_resource_serialization_poem_serde_equivalence() {
-        let entry = PublicOplogEntry::CreateResource(ResourceParameters {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-            id: WorkerResourceId(100),
-        });
-
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn drop_resource_serialization_poem_serde_equivalence() {
-        let entry = PublicOplogEntry::DropResource(ResourceParameters {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-            id: WorkerResourceId(100),
-        });
-
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn describe_resource_serialization_poem_serde_equivalence() {
-        let entry = PublicOplogEntry::DescribeResource(DescribeResourceParameters {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-            id: WorkerResourceId(100),
-            resource_name: "test".to_string(),
-            resource_params: vec![
-                ValueAndType {
-                    value: Value::String("test".to_string()),
-                    typ: str(),
-                },
-                ValueAndType {
-                    value: Value::Record(vec![Value::S16(1), Value::S16(-1)]),
-                    typ: record(vec![field("x", s16()), field("y", s16())]),
-                },
-            ],
-        });
-
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn log_serialization_poem_serde_equivalence() {
-        let entry = PublicOplogEntry::Log(LogParameters {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-            level: LogLevel::Stderr,
-            context: "test".to_string(),
-            message: "test".to_string(),
-        });
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
-    }
-
-    #[test]
-    #[cfg(feature = "poem")]
-    fn restart_serialization_poem_serde_equivalence() {
-        let entry = PublicOplogEntry::Restart(TimestampParameter {
-            timestamp: rounded_ts(Timestamp::now_utc()),
-        });
-        let serialized = entry.to_json_string();
-        let deserialized: PublicOplogEntry = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(entry, deserialized);
     }
 }
